pid works motion profile blitz

+++ pico/pio_pico/lib/Blitz/serial_interfaces.hpp
@@ -19,7 +19,7 @@ struct PIDFeedback {
     uint8_t id;
     float setpoint;
     float current;
-    float motor_pwm;
+    int16_t motor_pwm;
     float elapsed_time;
 
 };
